pyproject.toml
[project]
name = "wikipedia-mcp-service"
version = "0.1.0"
description = "Wikipedia MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "google-adk>=1.0.0",
    "google-genai==1.38.0",
    "google-cloud-storage>=2.19.0",
    "networkx>=3.5",
    "pydantic>=2.11.7",
    "python-dotenv==1.1.0",
    "thefuzz==0.22.1",
    "python-Levenshtein==0.25.1",
    "floggit>=0.0.19",
]
